Change reset_patience from int to float

+++ ccreaim/model/vqvae.py
@@ -32,7 +32,7 @@ class VQVAE(nn.Module):
 
 
 class VectorQuantizer(nn.Module):
-    def __init__(self, num_embeddings: int, embedding_dim: int, reset_patience: int):
+    def __init__(self, num_embeddings: int, embedding_dim: int, reset_patience: float):
         super().__init__()
         self.K = num_embeddings
         self.D = embedding_dim
